bugfix for merging blocks during 'delete' keypress & bugfix for console errors
If you add a header text with a source <h1>test</h1> and in the middle of this text (for example after "e" letter) press "enter" key. After that, move your caret to the end of the first block. Press "delete" key. You will see this bug occuring. 

This is a bugfix for similar cases.